Fix loadFile for edge case machines (#2)
On some machines just loading the current path as a string won't work, for instance, there are some users on Windows who have a "#" inside there username.  This will cause loading the path directly to break, as everything after the # will be treated as an anchor.  The solution is to use `url.format` which will escape URL components appropriately.